WIP: Refactor Modals
