+++ src/common/Backdrop/index.js
@@ -1,3 +1,5 @@
+import { motion } from 'framer-motion';
+
 import styles from './index.module.scss';
 
 const Backdrop = ({ toggleModal }) => {
@@ -7,7 +9,17 @@ const Backdrop = ({ toggleModal }) => {
     }
   };
 
-  return <div className={styles.backdrop} onClick={handleClick}></div>;
+  return (
+    <motion.div
+      key="backdrop"
+      className={styles.backdrop}
+      onClick={handleClick}
+      initial={{ opacity: 0 }}
+      animate={{ opacity: 1 }}
+      transition={{ duration: 0.1 }}
+      exit={{ opacity: 0 }}
+    />
+  );
 };
 
 export default Backdrop;
