+++ src/common/CenterModal/index.js
@@ -1,27 +1,71 @@
 import { createPortal } from 'react-dom';
 
-import { useKeyDown } from 'hooks/useKeyDown';
+import { motion, AnimatePresence } from 'framer-motion';
+import { useMediaQuery } from 'react-responsive';
+
+// import { useKeyDown } from 'hooks/useKeyDown';
 
 import Backdrop from 'common/Backdrop';
 
 import styles from './index.module.scss';
 
-const CenterModal = ({ children, toggleModal, className }) => {
+const CenterModal = ({
+  children,
+  toggleModal,
+  backdropClassName,
+  modalClassName,
+}) => {
   const backdropElement = document.getElementById('backdrop');
   const overlaysElement = document.getElementById('overlays');
 
-  useKeyDown(() => {
-    toggleModal();
-  }, ['Escape']);
+  // useKeyDown(() => {
+  //   toggleModal();
+  // }, ['Escape']);
+
+  const isBigScreen = useMediaQuery({
+    query: '(min-width: 768px)',
+  });
+
+  const variants = isBigScreen
+    ? {
+        initial: { x: '50%', y: '100vh' },
+        visible: { x: '50%', y: '-50%' },
+        exit: { x: '50%', y: '100vh' },
+      }
+    : {
+        initial: { x: '50%', y: '100vh' },
+        visible: { x: '50%', y: 0 },
+        exit: { x: '50%', y: '100vh' },
+      };
 
   return (
-    <>
-      {createPortal(<Backdrop toggleModal={toggleModal} />, backdropElement)}
-      {createPortal(
-        <div className={`${styles.modal} ${className}`}>{children}</div>,
-        overlaysElement
+    <AnimatePresence>
+      {children && (
+        <>
+          {createPortal(
+            <Backdrop
+              toggleModal={toggleModal}
+              className={backdropClassName}
+            />,
+            backdropElement
+          )}
+          {createPortal(
+            <motion.div
+              className={`${styles.modal} ${modalClassName}`}
+              key="centermodal"
+              variants={variants}
+              initial="initial"
+              animate="visible"
+              transition={{ duration: 0.2 }}
+              exit="exit"
+            >
+              {children}
+            </motion.div>,
+            overlaysElement
+          )}
+        </>
       )}
-    </>
+    </AnimatePresence>
   );
 };
