+++ src/common/SideModal/index.module.scss
@@ -1,6 +1,7 @@
 @import 'App.scss';
 
 .modal {
+  overflow-y: hidden;
   position: fixed;
   right: 50%;
   transform: translate(50%, 0);
