+++ src/components/CartContent/index.module.scss
@@ -1,11 +1,5 @@
 @import 'App.scss';
 
-.modal {
-  @media (min-width: 900px) {
-    overflow-y: hidden;
-  }
-}
-
 .empty {
   width: 95%;
   margin: 0 auto;
