+++ src/pages/Account/AccountProfile/index.js
@@ -2,7 +2,9 @@ import { useState } from 'react';
 
 import { BiUser, BiEnvelope, BiPhone } from 'react-icons/bi';
 
-import EditProfileModal from './EditProfileModal';
+import EditProfile from './EditProfile';
+
+import CenterModal from 'common/CenterModal';
 
 import styles from './index.module.scss';
 
@@ -15,14 +17,20 @@ const AccountProfile = ({ name, email, lastName, phoneNumber }) => {
 
   return (
     <>
-      {isOpen && (
-        <EditProfileModal
-          toggleEditProfile={toggleEditProfile}
-          name={name}
-          lastName={lastName}
-          phoneNumber={phoneNumber}
-        />
-      )}
+      <CenterModal
+        modalClassName={styles.modal}
+        toggleModal={toggleEditProfile}
+      >
+        {isOpen && (
+          <EditProfile
+            toggleEditProfile={toggleEditProfile}
+            name={name}
+            lastName={lastName}
+            phoneNumber={phoneNumber}
+          />
+        )}
+      </CenterModal>
+
       <div className={styles.profile_container}>
         <div className={styles.profile_wrapper}>
           <h3 className={styles.profile_title}>Mis datos</h3>
