--- src/pages/Addresses/AddAddressModal/index.module.scss
@@ -1,110 +1,0 @@
-@import 'App.scss';
-
-.modal {
-  padding-right: 3rem;
-  padding-left: 3rem;
-
-  .form {
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-
-    .title {
-      text-transform: uppercase;
-      text-align: center;
-      margin-bottom: 2rem;
-    }
-
-    .form_inputs_wrapper {
-      overflow-y: auto;
-
-      &::-webkit-scrollbar {
-        width: 1rem;
-      }
-
-      &::-webkit-scrollbar-track {
-        background: hsl(0, 0%, 95%);
-        border-radius: 100vw;
-      }
-
-      &::-webkit-scrollbar-thumb {
-        background: hsl(0, 0%, 76%);
-        border-radius: 100vw;
-      }
-
-      &::-webkit-scrollbar-thumb:hover {
-        background: hsl(0, 0%, 66%);
-      }
-
-      .label {
-        span {
-          font-family: $font-f-secondary;
-          font-weight: $font-w-700;
-          letter-spacing: $ls-secondary;
-        }
-
-        .input {
-          border: solid 0.2rem #9e9e9e;
-          border-radius: 1rem;
-          padding: 1rem 2rem;
-          width: 100%;
-          font-size: 1.6rem;
-          margin-bottom: 3rem;
-        }
-      }
-
-      .checkbox {
-        display: flex;
-        line-height: normal;
-        font-family: $font-f-secondary;
-        font-weight: $font-w-700;
-        letter-spacing: $ls-secondary;
-
-        input {
-          width: auto;
-          margin-left: 1rem;
-          margin-right: 1rem;
-        }
-      }
-    }
-
-    .button_wrapper {
-      padding-top: 2rem;
-      margin-top: auto;
-
-      .button {
-        padding: 1rem 2rem;
-        font-size: 1.6rem;
-        border: solid 0.2rem $back-c-button-primary;
-        border-radius: 1rem;
-        width: 100%;
-        // override tailwind css
-        background: $back-c-button-primary !important;
-        color: white;
-        text-transform: uppercase;
-        letter-spacing: $ls-button;
-
-        -webkit-user-select: none;
-        -webkit-touch-callout: none;
-        -moz-user-select: none;
-        -ms-user-select: none;
-        user-select: none;
-
-        -moz-transition: all 0.1s ease-in;
-      -o-transition: all 0.1s ease-in;
-      -webkit-transition: all 0.1s ease-in;
-      transition: all 0.1s ease-in;
-
-        &:active {
-          transform: scale(0.9);
-        }
-
-        &:hover {
-          cursor: pointer;
-          background: $back-c-button-hover !important;
-          border-color: $back-c-button-hover;
-        }
-      }
-    }
-  }
-}
