+++ src/pages/Addresses/Address/index.module.scss
@@ -1,5 +1,10 @@
 @import 'App.scss';
 
+.modal {
+  padding-right: 3rem;
+  padding-left: 3rem;
+}
+
 .card {
   display: flex;
   flex-direction: column;
