+++ src/pages/Addresses/index.module.scss
@@ -1,5 +1,10 @@
 @import 'App.scss';
 
+.modal {
+  padding-right: 3rem;
+  padding-left: 3rem;
+}
+
 .container {
   padding: 10rem 0 4rem 0;
