+++ src/setup/Layout/Header/SideNav/index.module.scss
@@ -1,9 +1,5 @@
 @import 'App.scss';
 
-.modal {
-  padding: 3rem 1rem 0;
-}
-
 .container {
   display: flex;
   flex-direction: column;
