+++ src/setup/Layout/Header/index.module.scss
@@ -1,0 +1,13 @@
+.backdrop {
+  @media (min-width: 900px) {
+    display: none;
+  }
+}
+
+.side_nav {
+  padding: 3rem 1rem 0;
+
+  @media (min-width: 900px) {
+    display: none;
+  }
+}
